Extracted out a Formatter class so that TextUI adheres to SRP
