+++ src/seedu/addressbook/Main.java
@@ -13,6 +13,7 @@ import seedu.addressbook.parser.Parser;
 import seedu.addressbook.storage.StorageFile;
 import seedu.addressbook.storage.StorageFile.InvalidStorageFilePathException;
 import seedu.addressbook.storage.StorageFile.StorageOperationException;
+import seedu.addressbook.ui.Formatter;
 import seedu.addressbook.ui.TextUi;
 
 
@@ -26,6 +27,7 @@ public class Main {
     public static final String VERSION = "AddressBook Level 2 - Version 1.0";
 
     private TextUi ui;
+    private Formatter formatter;
     private StorageFile storage;
     private AddressBook addressBook;
 
@@ -52,13 +54,14 @@ public class Main {
      */
     private void start(String[] launchArgs) {
         try {
-            this.ui = new TextUi();
+            this.formatter = new Formatter();
+            this.ui = new TextUi(formatter);
             this.storage = initializeStorage(launchArgs);
             this.addressBook = storage.load();
-            ui.showWelcomeMessage(VERSION, storage.getPath());
+            formatter.showWelcomeMessage(VERSION, storage.getPath());
 
         } catch (InvalidStorageFilePathException | StorageOperationException e) {
-            ui.showInitFailedMessage();
+            formatter.showInitFailedMessage();
             /*
              * ==============NOTE TO STUDENTS=========================================================================
              * We are throwing a RuntimeException which is an 'unchecked' exception. Unchecked exceptions do not need
@@ -74,7 +77,7 @@ public class Main {
 
     /** Prints the Goodbye message and exits. */
     private void exit() {
-        ui.showGoodbyeMessage();
+        formatter.showGoodbyeMessage();
         System.exit(0);
     }
 
@@ -86,7 +89,7 @@ public class Main {
             command = new Parser().parseCommand(userCommandText);
             CommandResult result = executeCommand(command);
             recordResult(result);
-            ui.showResultToUser(result);
+            formatter.showResultToUser(result);
 
         } while (!ExitCommand.isExit(command));
     }
@@ -112,7 +115,7 @@ public class Main {
             storage.save(addressBook);
             return result;
         } catch (Exception e) {
-            ui.showToUser(e.getMessage());
+            formatter.showToUser(e.getMessage());
             throw new RuntimeException(e);
         }
     }
